+++ app/tests/utils/env.py
@@ -20,7 +20,7 @@ TEST_BCC_PORT = 8000
 
 TEST_MSS_APP_TOKEN = "some-mss-app-token-for-testing"
 
-TEST_EXECUTOR_CONFIG_FILE = get_fixture_path("dummy-executor-config.yml")
+TEST_QUANTIFY_CONFIG_FILE = get_fixture_path("dummy-quantify-config.yml")
 TEST_BACKEND_SETTINGS_FILE = get_fixture_path("backend_config.toml")
 TEST_SIMQ1_BACKEND_SETTINGS_FILE = get_fixture_path("backend_config.simq1.toml")
 
@@ -49,5 +49,5 @@ def setup_test_env():
     environ["BCC_PORT"] = f"{TEST_BCC_PORT}"
 
     environ["MSS_APP_TOKEN"] = TEST_MSS_APP_TOKEN
-    environ["EXECUTOR_CONFIG_FILE"] = TEST_EXECUTOR_CONFIG_FILE
+    environ["QUANTIFY_CONFIG_FILE"] = TEST_QUANTIFY_CONFIG_FILE
     environ["IS_AUTH_ENABLED"] = "True"